Tests/LibWeb/Text/expected/WebAudio/AudioWorklet-registerProcessor-parameterDescriptors-validation.txt

duplicate names: NotSupportedError
defaultValue outside range: InvalidStateError
PASS
